Updated the parsers Makefile.am to actually make use of the results returned by the autoconf/automake magic.
The code compiles in Linux and FreeBSD now, though using the 'make dist' tarball in FreeBSD requires the *.tab.* files to be deleted. These are created by bison in Linux and are meant to be part of the dist so that whoever uses it doesn't need bison themselves...they dont seem to work nicely in FreeBSD for some reason?