Author: Tom Young
Here is a patch to use the control message block, rather than the ioctl for
getting the time stamp. It seems to work for me. As far as i can tell, the
control block stuff has been used for ages (it looked like it was in 2.4.24)
but it should fall back to the ioctl anyway if it has to.