Changed AC_PROG_YACC test for AC_CHECK_PROGS, as AC_PROG_YACC assumes "yacc" is always installed
Added checks to see if $YACC and $LEXLIB are populated. If not, generate a warning saying the tools aren't installed but are required, and increment and error counter. Finally, check the errorcounter and generate a critical error if nonzero.

Test $LEXLIB because it will be empty if there's no lex installed - $LEX gets filled with a autoconf "missing" command instead, which simply results in deferred failure (eg, at the point it tries to use lex)