start of a trace replay tool.
The tool is supposed to play back a trace in trace time, using the libtrace events to manage the actual timings. The tool is based on the event and output examples.

This is the very start of the tool, and it currently segfaults whenever it is run. 
The order of business is to: 
1. Make it work.]
2. Refactor it so that work is done _before_ the next packet needs to be sent, unlike now. This will lead to less missed deadlines.
3. Investigate if it is possible to use a resulotion wait function for greater accuracy.