Added a new format: rawerf. This is a special format that will force ERF files to be opened and read without going through the compression autodetection. It will only work for uncompressed ERF files.
This should be a handy option for people who are working with lots of uncompressed ERF files -- running the compression autodetection against uncompressed ERF traces can very occasionally produce a false positive, i.e. decide the file is compressed. By skipping the autodetection and forcing the traces to be read as uncompressed ERF, the threat of this bug will go away.