Adds tests for the Live Linux formats, int: ring: and pcapint:
Launch from test/do-live-tests.sh (needs root)
Alternatively test/netns-env <command>, bash can be used here useful for debugging etc.

This runs tests in separate network namespace with two virtual adapters, this is isolated from the kernels influence so that the only traffic is that which we create.
* test-live.c - Tests sending and receiving 100 packets from one format to another between to virtual ports
                               * Tests read/write, wirelength, capturelength, packet order/integrity, timestamps, ability to set direction/snap lengths  and trace statistics
* test-live-snaplen.c Tests that snap length works correctly via trace_config(), does basic checks on the wire and capture lengths returned

Limitations
* Doesn't support BSD yet
* DPDK included but probably not working in master
* Doesn't test BPF filters
* Requires root to run

Found some existing issues :(, fixes to which will be in the following commits.