Better detection of truncation in libwandio zlib
Previously, we were returning "success" when reading from a truncated
gzip file and simply returning all bytes that we could successfully
decompress. This is not consistent with other gzip tools, which at least
alert that the file was not complete or the gzip integrity check was missing.

Now we will return an error if we hit an EOF in the gzip stream that is
not immediately preceded by a Z_STREAM_END marker and write an error message
to stderr so the user can know what has occurred.