Fix dodgy stream buffer flushing in DAG
Not sure what was going on with that flushing code, but I'm pretty sure
that wasn't going to work.

Ignore loss counters on the first packet observed on each interface.
Modern DAG interfaces treat any packets that appear while you're not
capturing as lost, so stopping and restarting a DAG capture will always
result in a non-zero loss counter. We tend to use the loss counter to
only indicate packets lost during active capture, so we need to compensate.