Code style fixes.
There are a couple of dubious lines in there that someone else might
like to change. Also, formatting relies on 8 space tabs which could be
improved by tabbing to the save level and spacing to line up with the
previous line when trying to break a function call across multiple lines.