Refactoring the packet loop
Removes the pread type selection function and replaces it with a
function pointer created when the format is started.

Tidies the per packet loop, removes duplicate code and rewrites
tracetime-delaying so that a message can be processed while waiting
for the delay to return within a batch of packets.

In this case if a pause message is encountered we will first
notify the trace that we are pausing. Then send any remaining
packets from the batch without delay, and then properly pause. The resume
message is sent to the per packet thread before normal packet
flow continues.

This still is WIP and contains alot of debugging code. The error
path still needs to be better implemented