Fix bug in ring: format if the interface is down
If the ring interface goes down (or starts down), Linux considers this
to be an error for the socket and we were not checking for the error
returns of the individual poll() items, where this error was returned.

Due to ordering of the code this broke libtrace_halt (if no packets are being received),
even if the interface came back (once shutdown the error seems to persist
upon restart).

The behaviour now is that an error, such as the interface being down
(probably the only error one could encounter?) will return as an error
to the libtrace application.