Add support to wandio for reading over HTTP
This adds a new IO Reader, HTTP which uses libcurl to read data from HTTP
(tested also with HTTPS).  The implementation is a lightly modified version of
the code used in the kurl library available at
https://github.com/attractivechaos/klib.

Regular reading has been tested (including processing trace files with
libtrace), but `seek` and `tell` have not been tested. I would be perfectly
happy if we remove this functionality, but I'll leave that up to you.

Known Issues:
 - When reading a trace from an HTTP URI, the trace format _must_ be
   specified, otherwise libtrace thinks that the format is `http`. It isn't
   clear to me the "correct" way to fix this (or even that it needs fixing.)
 - Calls to `http_open` and `http_close` are not strictly thread-safe because
   they call `curl_global_init`, which itself is not thread-safe. I have
   attempted to mitigate this problem by wrapping the call to `curl_global_init`
   in a mutex, but if users have other threads that are independently making
   calls to the same non-thread-safe functions that curl is...
 - OpenSSL support is almost certainly not thread-safe
   (http://curl.haxx.se/mail/lib-2004-08/0033.html).
 - When reading _very_ slowly over HTTP, the server may terminate the connection
   based on an inter-read timeout (`send_timeout` in NGINX).