Fixes an issue with old programs reusing a packet over multi traces
Fixes a bug that Shane found. In the case an old libtrace program used
a single packet and cycled over multiple traces. Closing the trace
and then reusing the packet with a new trace will try to finish
the packet which requires the old format to be present.

As such we now keep track of the last packet against the format,
and finish this when the trace is destroyed.