Fix bad framing calculation when converting to ERF
The bug was caused by our newly added support for ERF extension
headers. Previously, erf_get_framing_length() didn't actually
inspect the contents of the packet passed to it but now it does (to
try and detect any extension headers).

This means that when converting from another format, we can't call
erf_get_framing_length on the original packet anymore to work out
how much framing space we need as it will interpret the original
packet contents as an ERF header.