Make RT safe for use as a parallel input
Previously, the RT format used to assume that only one read packet would
be in use at any given time, so subsequent read_packet calls could overwrite
the buffer space that previous packets had occupied. With parallel libtrace,
this is no longer the case.

For now, I've gone with the safe but slow solution of copying every
complete RT packet out of the receive buffer into memory space that has
been allocated to the libtrace_packet_t. It'd be nice to do this without
the copy, but we'd need to start tracking which packets have been returned
to us to be able to do this properly.