Add bucket data structure for keeping track of packet buffers
A bucket can contain anywhere from 1 to N packets and will point to a
buffer owned by libtrace. Libtrace can signal that it is finished with an
individual packet and when all packets associated with the buffer are
finished, the buffer will be freed automatically.

This should make buffer management easier for formats like bpf and RT where
multiple packets can be read at once and will allow us to retain zero-copy
behaviour (aside from the initial read) in a parallel libtrace program.