Fix deadlock when reading parallel RT with a filter
The problem was that filtered packets were still consuming slots in the
buckets, but matching packets would only get released once parallel libtrace had
read a full batch (10 packets, by default). If the filter didn't match many
packets, the 200,000 packet slots would wrap around before a batch was
complete -- leaving us with nowhere to store new packets.

The fix was to allow the bucket structure to "reclaim" a slot if it is
released as soon as it is allocated, i.e. if the packet was filtered. This
means that only matching packets will consume slots and we can't run out
before we read a full batch.