Add -N option to tracertstats
This will set the trace burst size to 1, so that libtrace will not block
waiting for a full packet buffer before processing packets or ticks. This
is useful when using an interface with very little traffic as input, as
otherwise the output is quite bursty rather than the intended "real time".