Add usleep for first packet in all parallel tests
This ensures that the processing and reporting threads will still be
running when our test tries to pause them. Without this, our tests will
sometimes fail because we call pause on a trace that is finished (i.e. not
started).

Started noticing this with gcc 5.1.1, so might be due to better compiler
optimisations?