Fix issues with bad ordering in the ordered combiner
Mostly this was down to a false assumption that interval ticks
would be processed at the same time by each processing thread, i.e. all
packets after a tick in one thread would always follow the packets
preceding the same tick in another thread.

Instead, we now simply skip past ticks in the queues when considering
what is next in the queue, rather than trying to continue looping until
all queues have the same tick as their next message.

Of course, we still pass the first instance of any given tick on to the
reporter as soon as we see it -- now we just continue trying to process
the queue it came from if the next result has the lowest order value.