Seed the RNG when calling trace_init
If we don't do this, we can't have more than one ring/int parallel input
at a time on a host. This is because the fanout group id is chosen using
rand() and so was previously always the same value, leading to conflicts
that would cause trace_pstart() to fail.