More pfring improvements
Only byteswap if the byte-ordering of the pfring header doesn't
match the byte ordering of the host looking at it. It's a bit of a pain,
but we were losing quite a bit of performance by always byteswapping to
a standard byte order upon reading the packet.

Turns out the pfring API has a get_selectable_fd function so our event
API doesn't need to poll if no packets are available.