Couple of optimisations introduced after looking at callgrind output
Only reset trace->last_packet to NULL in situations where we aren't
about to immediately read a new packet. The overhead of grabbing
and releasing the mutex to do this is massive for an operation we
are about to immediately overwrite.

Removed loop that checks the integrity of the packet array before
filling it with new data. This consumes a lot of CPU cycles for a
condition that really shouldn't happen anyway -- the fact that we
have asserts here suggest that this is more of a sanity check during
development than a condition we might have to deal with.