Fix segfault when destroying a packet from an EOF'd trace.
The problem is caused by not setting packet->trace to be NULL when
read_packet is going to return EOF. As a result, the trace_destroy()
code thinks that there are no packets pointing to the trace about
to be destroyed. Since there are no packets that need to be unlinked
from the trace (i.e. have their ->trace pointers set to NULL), the
unlinking step is skipped.

In actuality, we still have the packet that we tried
to read into when we got an EOF still pointing at the trace -- as soon
as we try to destroy that packet, it tries to access the trace structure
that we just freed with trace_destroy(), causing the segfault.