Fix for issue #39 - ring and int pstop() fails on older kernels when using threads
The problem here is that on old kernels without PACKET_FANOUT support
(added in v3.1) will only include the single threaded versions of int
and ring. When used with multiple threads the libtrace API will
fallback to using read rather than pread which does not check message
queues.

To fix this issue, in any format without pread support:
* We check for new messages with each loop around read_packet as we fill the burst
* Within read_packet we update the halt to include the pausing state
* Use a seperate lock to the main lock when reading a burst of packets,
  otherwise trace_ppause has to wait for a burst to read.

This is not 100% perfect as a single packet might still need to be received
before a generic message can be received.
A proper fix in the future would be to move all format internals purely to the
parallel API.