Remove mutex in the packet path, which limits parallel performance
Packets should be destroyed (or finished) before the format is closed,
however this was never enforced or implemented in the single threaded API.

As such to provide support for old applications that destroy the packet
after the trace we store the packet against the trace. This was a race
and locks were used for safety even when using the parallel API.

Instead of locks we can assume that the parallel API will destroy
packets before the trace is closed. So instead of locks only store
the packet in the single threaded case.