Fix incorrectly reported test failures in test-live
test-live was not taking the smaller test-size into account when
evaluating whether tests where pcapint was the source had
succeeded. Therefore they would report as failed because the
other end had only received 30 packets instead of 100, but
the pcapint source was set to only send 30 packets in the first
place!