Update test suite for pcapng
Also added a new test for set_capture_length(). Setting the caplen
for pcapng is a bit more complicated than most formats, so I
needed a test for that format anyway. Figured I may as well add
tests for some of the other truncatable formats as well.

I've also added a 'complex.pcapng' test trace -- I haven't written
any tests that use it but it is useful for checking if you can
still read the full variety of pcapng records.