Add new format for receiving and parsing nDAG records
Also added new statistic: missing_records -- this tracks records
that have gone missing between the original capture device and
the current processing host (i.e. lost on the network between the
two). Useful for nDAG as there is a distinction between packets
dropped by the DAG and packets that were not received by the nDAG
client.