Fix bad IPv6 fragment offset calculation.
This has been horribly broken the whole time, but we generally got
lucky and fell through to the default "no fragmentation" case.
Since nobody uses IPv6 fragmentation, this seemed to be giving the
correct result.

Of course, on the odd occasion we'd accidentally treat a packet as
a fragment incorrectly and therefore trace_get_source_port() etc
would fail -- this was a bit more problematic :)