Only initialise mmsgbufs when we're actually going to use them.
Previous change meant that we aren't going to be actually calling
recvmmsg so stop wasting time repeatedly initialising buffers that
aren't going to be used.

Also moved initialisation of mmsgbuf fields that do not change
out of the per-packet path so we don't end up repeatedly setting
values to the same thing that they were before.