Avoid multiple byteswaps of the same variable in format_dag.
In particular, we can cache the timestamp of the next packet
in a socket buffer rather than re-read it from the buffer every
time we call select_next_packet() -- maybe not such a big issue
when you have two sockets per thread, but very wasteful when you
have 8 sockets per thread.

Keep a local copy of rlen in host byte order when preparing a
packet, so we don't have to repeatedly swap it. Compiler
optimization might make this improvement irrelevant, though?