Add DLLEXPORT to all hash_toeplitz API functions.
This allows the toeplitz hasher to be used outside of
libtrace -- e.g. a user can create their own hasher and pass
it into trace_set_hasher.

This is useful if a user wants to use the same hasher (and
therefore the same seed) across multiple traces -- libtrace
in-built hashers are re-seeded for each trace file so their
hashing is not consistent across multiple inputs.