Fixes trace_pstop not working with pcapint issue #69
Fixes a deadlock which was caused by allowing a lock in a single threaded part of the code.
This would clear itself if enough packets are received, but not on a quiet link.

This lock was originally added to fix a bug I introduced allowing multiple
threads into that code. Introduced in 47c4490, which has since been fixed
properly by 684a21c.

If we need to lock on a read operation there is a read_lock which should be used
instead of the trace_lock.