Refactor pcapng to read an entire block when possible
Remove much of the duplicated read logic in pcapng. Most blocks were
processed by reading the header, and then reading the variable length
portions such as a packet or options.

Instead, this is all replaced with a single top level read which reads
the entire block. This simplifies logic and makes it easier to check
for out of bounds reads/writes. This also adds additional
sanity checks, such as checking the caplen actually matches with the
data supplied.

The one exception to this refactor are 'section' blocks which contain the
magic to determine endianness. So we don't know the endianess of a block
length until we read a 'section'.