Fix broken payload lengths when IP length is not yet set.
Presumably, this can happen when the packet segmentation has
been offloaded to the NIC hardware -- the length is not set
until the packet is about to be sent (i.e. *after* an outgoing
packet is captured), so our calculation ends up producing a
negative number (0 - sizeof(IP) - sizeof(TCP / UDP)).

In this case, we instead have to infer the payload length using
the amount of bytes captured less the difference between the start
of the IP header and the start of the captured frame.