Clear error in trace_set_hasher() before resolving the problem
If the format-specific configuration function fails when setting
the hasher, we fallback to trying to use a standard built-in
hasher. If this is successful, we need to make sure that the
trace error state is cleared otherwise the caller could call
trace_is_err() on the trace and think there was an error when
really we have already resolved the issue.