Sneaky attempt to avoid depending on dpdk-dev for jessie
Jessie doesn't have dpdk-dev -- previously we were relying on
the packages in the WAND repo that were built for OVS, but that
repo is going away and I'm disinclined to try and port all of
the DPDK packages over to the WAND bintray right now (there's
like 30 different packages at least).

This does mean that any jessie libtrace packages will not
support DPDK from 4.0.6 forward -- this *probably* won't be an
issue for most users (?) and will eventually resolve itself once
jessie drops off the list of supported suites.