trace_post_reporter(): fully initialize message
Valgrind complains that a call to write has a buffer that isn't fully
initialized. This is due to a possibility of the uninitialized bytes in
that structure could be sensitive memory from its previous contents.

This patch makes sure that the full libtrace_message_t gets initialized
with 0s and then proceeds to set the contents before moving on to
calling the function that calls write().