Improve parallel performance by skipping needless sanity checks
Normally, the sanity checking that we do here would be a good idea,
just in case something has corrupted one of our packets.

BUT, these checks happen every single time we attempt to read a
batch of packets (even for members of the packet array that are
not filled in by the read operation) and even though the checks
themselves are pretty simple, the CPU cycles used to perform them
really start to add up at high packet rates.

Instead, we're going to trust that a) we don't invalidate the
packets ourselves with some broken internal code and b) the user
doesn't go around touching raw packet structure contents
because we have an API for all necessary interactions. Sanity
checks can also be added at the point where the packet array element
is about to be updated by the read, if necessary. At least at that
point, we're sanity-checking something that we're about to use
instead of the current case where we waste time checking something
that we may or may not use.